Move `PhysicalItemPF2e#generateModifiedName` to globally-exposed helper
The name-generation method doesn't work very well (or at all) for some languages. This will allow localization-module
authors to replace the function with their own.
