+++ src/module/item/physical/document.ts
@@ -19,7 +19,6 @@ import {
     Price,
 } from "./data.ts";
 import { CoinsPF2e, computeLevelRarityPrice } from "./helpers.ts";
-import { RUNE_DATA } from "./runes.ts";
 import { getUsageDetails, isEquipped } from "./usage.ts";
 import { DENOMINATIONS } from "./values.ts";
 
@@ -261,7 +260,7 @@ abstract class PhysicalItemPF2e<TParent extends ActorPF2e | null = ActorPF2e | n
     override prepareDerivedData(): void {
         super.prepareDerivedData();
 
-        this.name = this.generateModifiedName();
+        this.name = game.pf2e.system.generateItemName(this);
 
         this.system.identification.identified ??= {
             name: this.name,
@@ -457,74 +456,6 @@ abstract class PhysicalItemPF2e<TParent extends ActorPF2e | null = ActorPF2e | n
             description: { value: this.description },
             material,
         };
-    }
-
-    /**
-     * Generate a modified item name based on precious materials and runes. Currently only armor and weapon documents
-     * have significant implementations.
-     */
-    generateModifiedName(): string {
-        if (!this.isOfType("armor", "weapon")) return this.name;
-
-        type Dictionaries = [
-            Record<string, string | undefined>,
-            Record<string, { name: string } | undefined>,
-            Record<string, { name: string } | null>
-        ];
-
-        // Acquire base-type and rune dictionaries, with "fundamental 2" being either resilient or striking
-        const [baseItemDictionary, propertyDictionary, fundamentalTwoDictionary]: Dictionaries = this.isOfType("armor")
-            ? [CONFIG.PF2E.baseArmorTypes, RUNE_DATA.armor.property, RUNE_DATA.armor.resilient]
-            : [CONFIG.PF2E.baseWeaponTypes, RUNE_DATA.weapon.property, RUNE_DATA.weapon.striking];
-
-        const storedName = this._source.name;
-        const baseType = this.baseType ?? "";
-        if (
-            !baseType ||
-            !(baseType in baseItemDictionary) ||
-            this.isSpecific ||
-            storedName !== game.i18n.localize(baseItemDictionary[baseType] ?? "")
-        ) {
-            return this.name;
-        }
-
-        const { material } = this;
-        const { runes } = this.system;
-        const potencyRune = runes.potency;
-        const fundamental2 = "resilient" in runes ? runes.resilient : runes.striking;
-
-        const params = {
-            base: baseType ? game.i18n.localize(baseItemDictionary[baseType] ?? "") : this.name,
-            material: material.type && game.i18n.localize(CONFIG.PF2E.preciousMaterials[material.type]),
-            potency: potencyRune,
-            fundamental2: game.i18n.localize(fundamentalTwoDictionary[fundamental2]?.name ?? "") || null,
-            property1: game.i18n.localize(propertyDictionary[runes.property[0]]?.name ?? "") || null,
-            property2: game.i18n.localize(propertyDictionary[runes.property[1]]?.name ?? "") || null,
-            property3: game.i18n.localize(propertyDictionary[runes.property[2]]?.name ?? "") || null,
-            property4: game.i18n.localize(propertyDictionary[runes.property[3]]?.name ?? "") || null,
-        };
-        // Construct a localization key from material and runes
-        const formatString = (() => {
-            const potency = params.potency && "Potency";
-            const fundamental2 = params.fundamental2 && "Fundamental2";
-            const properties = params.property4
-                ? "FourProperties"
-                : params.property3
-                ? "ThreeProperties"
-                : params.property2
-                ? "TwoProperties"
-                : params.property1
-                ? "OneProperty"
-                : null;
-            const material = params.material && "Material";
-            const key =
-                [potency, fundamental2, properties, material]
-                    .filter((keyPart): keyPart is string => !!keyPart)
-                    .join("") || null;
-            return key && game.i18n.localize(key);
-        })();
-
-        return formatString ? game.i18n.format(`PF2E.Item.Physical.GeneratedName.${formatString}`, params) : this.name;
     }
 
     async setIdentificationStatus(status: IdentificationStatus): Promise<void> {
