+++ src/scripts/set-game-pf2e.ts
@@ -2,7 +2,7 @@ import { Action } from "@actor/actions/index.ts";
 import { AutomaticBonusProgression } from "@actor/character/automatic-bonus-progression.ts";
 import { ElementalBlast } from "@actor/character/elemental-blast.ts";
 import { CheckModifier, ModifierPF2e, StatisticModifier } from "@actor/modifiers.ts";
-import { CoinsPF2e } from "@item/physical/helpers.ts";
+import { CoinsPF2e, generateItemName } from "@item/physical/helpers.ts";
 import { CompendiumBrowser } from "@module/apps/compendium-browser/index.ts";
 import { EffectsPanel } from "@module/apps/effects-panel.ts";
 import { LicenseViewer } from "@module/apps/license-viewer/app.ts";
@@ -93,7 +93,7 @@ export const SetGamePF2e = {
             licenseViewer: new LicenseViewer(),
             rollActionMacro,
             rollItemMacro,
-            system: { moduleArt: new ModuleArt(), remigrate, sluggify },
+            system: { generateItemName, moduleArt: new ModuleArt(), remigrate, sluggify },
             variantRules: { AutomaticBonusProgression },
         };
         game.pf2e = mergeObject(game.pf2e ?? {}, initSafe);
